\name{DJI_return}
\alias{DJI_return}
\docType{data}
\title{
Dow Jones Industrial Average (DJIA)
}
\description{
Dow Jones Industrial Average (DJIA) is a stock market index that shows how 30 large publicly owned companies based in the United States have traded during a standard NYSE trading session. We consider monthly cross-sectional returns from April 2004 to December 2017. The data were obtained from the CRSP (Center for Research in Security Prices) database.
}
\usage{data("DJI_return")}
\format{
A data matrix
}
\references{
Kokoszka, P., Miao, H., Petersen, A. and Shang, H. L. (2019) `Forecasting of density functions with an application to cross-sectional and intraday returns', \emph{International Journal of Forecasting}, \bold{35}(4), 1304-1317.
}
\examples{
data(DJI_return)
}
\keyword{datasets}
