\name{quantile}
\alias{quantile}
\title{Quantile}
\usage{
quantile(x, ...)
}

\arguments{
\item{x}{Numeric vector whose sample quantiles are wanted, or an object of a class for which a method has been defined.}
\item{...}{Arguments passed to specific methods.}
}

\description{Generic functions for quantile.}

\value{Refer to specific methods. For numeric vectors, see the \code{\link[stats]{quantile}} functions in the \code{stats} package.}

\author{Han Lin Shang}

\seealso{\code{\link[ftsa]{quantile.fts}}}
\keyword{models}