\name{sd}
\alias{sd}
\alias{sd.default}
\title{Standard deviation}
\usage{
sd(...)
}

\arguments{
\item{...}{Arguments passed to specific methods.}
}

\description{Generic functions for standard deviation.}

\details{The \code{\link[stats]{sd}} functions in the \code{stats} package are replaced by \code{sd.default}.}

\value{Refer to specific methods. For numeric vectors, see the \code{\link[stats]{sd}} functions in the \code{stats}
package.}

\author{Han Lin Shang}

\seealso{\code{\link[ftsa]{sd.fts}}}
\keyword{models}
